% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CisMRBEE_UV.R
\name{CisMRBEE_UV}
\alias{CisMRBEE_UV}
\title{CisMRBEE_UV: Cis Univariable Mendelian Randomization using Bias-corrected Estimating Equations}
\usage{
CisMRBEE_UV(
  by,
  bX,
  byse,
  bXse,
  LD,
  Rxy,
  xQTL.N,
  xQTL.selection.rule = "top_K",
  top_K = 1,
  xQTL.pip.min = 0.2,
  xQTL.max.L = 10,
  xQTL.cred.thres = 0.95,
  xQTL.pip.thres = 0.5,
  reliability.thres = 0.75,
  tauvec = seq(3, 30, by = 1.5),
  admm.rho = 2,
  coverage.xQTL = 0.95,
  coverage.causal = 0.95,
  max.iter = 100,
  max.eps = 0.001,
  ebic.gamma = 2,
  xQTLfit = NULL
)
}
\arguments{
\item{by}{A vector of effect estimates from the outcome GWAS.}

\item{bX}{A vector of effect estimates from the exposure GWAS.}

\item{byse}{A vector of standard errors of effect estimates from the outcome GWAS.}

\item{bXse}{A vector of standard errors of effect estimates from the exposure GWAS.}

\item{LD}{The LD matrix of variants.}

\item{Rxy}{The correlation matrix of estimation errors of exposures and outcome GWAS. The last column corresponds to the outcome.}

\item{xQTL.N}{The sample sizes of exposure.}

\item{xQTL.selection.rule}{The method for purifying informative xQTLs within each credible set. Options include "minimum_pip", which selects all variables with PIPs exceeding a specified threshold, and "top_K", which ensures at least K variables are selected based on their PIP ranking. Defaults to "top_K".}

\item{top_K}{The maximum number of variables selected in each credible sets. Defaults to 1.}

\item{xQTL.pip.min}{The minimum empirical PIP used in purifying variables in each credible set. Defaults to \code{0.2}.}

\item{xQTL.max.L}{The maximum number of L in estimating the xQTL effects. Defaults to 10.}

\item{xQTL.cred.thres}{The minimum empirical posterior inclusion probability (PIP) used in getting credible sets of xQTL selection. Defaults to \code{0.95}.}

\item{xQTL.pip.thres}{If SuSiE fails to find any credible set, the threshold of individual PIP when selecting xQTL. Defaults to \code{0.5}.}

\item{reliability.thres}{A threshold for the minimum value of the reliability ratio. If the original reliability ratio is less than this threshold, only part of the estimation error is removed so that the working reliability ratio equals this threshold.}

\item{tauvec}{A vector of tuning parameters used in penalizing the direct causal effect. Default is `seq(3,10,by=1)`.}

\item{admm.rho}{A parameter set in the ADMM algorithm. Default is \code{2}.}

\item{coverage.xQTL}{The coverage of defining a credible set in xQTL selection. Defaults to \code{0.95}.}

\item{coverage.causal}{The coverage of defining a credible set in cis-MRBEE. Defaults to \code{0.95}.}

\item{max.iter}{The maximum number of iterations for the ADMM algorithm. Default is \code{15}.}

\item{max.eps}{The convergence tolerance for the ADMM algorithm. Default is \code{0.005}.}

\item{ebic.gamma}{The extended BIC factor for model selection. Default is \code{2}.}

\item{xQTLfit}{Initial fits of xQTLs for exposures. This should only be yielded by SuSiE, as CARMA is not allowed for cis-UVMR analysis currently. Default is \code{NULL}.}
}
\value{
A list containing:
\describe{
\item{\code{theta}}{The estimated effect size of the tissue-gene pair.}
\item{\code{gamma}}{The estimated effect sizes of the direct causal variants.}
\item{\code{theta.cov}}{The variance of the estimated effect size `theta`.}
\item{\code{theta.se}}{The standard error of the estimated effect size `theta`.}
\item{\code{theta.z}}{The z-score of the estimated effect size `theta`.}
\item{\code{Bic}}{The BIC values for each tuning parameter.}
\item{\code{eQTL.fit}}{The SuSiE result of xQLT selection of exposure.}
\item{\code{var.error}}{The variance of residuals.}
\item{\code{var.error}}{The variance of infinitesimal effect.}
\item{\code{causal.fit}}{The SuSiE result of causal effect calibration of exposure.}
\item{\code{reliability.adjust}}{Estimated reliability-adjusted values.}
}
}
\description{
This function performs univariable cis-Mendelian randomization that removes weak instrument bias using Bias-corrected Estimating Equations and identifies uncorrelated horizontal pleiotropy (UHP).
}
