% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CisMRBEEX.R
\name{CisMRBEEX}
\alias{CisMRBEEX}
\title{Cis Multivariable Mendelian Randomization using Bias-corrected Estimating Equations}
\usage{
CisMRBEEX(
  by,
  bX,
  byse,
  bXse,
  LD,
  Rxy,
  model.infinitesimal = F,
  reliability.thres = 0.75,
  Lvec = c(1:5),
  causal.pip.thres = 0.2,
  xQTL.max.L = 10,
  xQTL.cred.thres = 0.95,
  xQTL.pip.thres = 0.5,
  xQTL.pip.min = 0.1,
  xQTL.Nvec,
  tauvec = seq(3, 30, by = 3),
  admm.rho = 2,
  ridge.diff = 1000,
  max.iter = 100,
  max.eps = 0.001,
  susie.iter = 500,
  ebic.theta = 1,
  ebic.gamma = 2,
  maxdiff = 3,
  theta.ini = F,
  gamma.ini = F,
  eQTLfitList = NULL
)
}
\arguments{
\item{by}{A vector of effect estimates from the outcome GWAS.}

\item{bX}{A matrix of effect estimates from the exposure GWAS.}

\item{byse}{A vector of standard errors of effect estimates from the outcome GWAS.}

\item{bXse}{A matrix of standard errors of effect estimates from the exposure GWAS.}

\item{LD}{The linkage disequilibrium (LD) matrix.}

\item{Rxy}{The correlation matrix of estimation errors of exposures and outcome GWAS. The last column corresponds to the outcome.}

\item{model.infinitesimal}{An indicator of whether using REML to model infinitesimal effects. Defaults to \code{F}.}

\item{reliability.thres}{A threshold for the minimum value of the reliability ratio. If the original reliability ratio is less than this threshold, only part of the estimation error is removed so that the working reliability ratio equals this threshold.}

\item{Lvec}{When SuSiE is used, the candidate vector for the number of single effects. Default is \code{c(1:min(10, nrow(bX)))}.}

\item{causal.pip.thres}{A threshold of minimum posterior inclusion probability. Default is \code{0.2}.}

\item{xQTL.max.L}{The maximum number of L in estimating the xQTL effects. Defaults to 10.}

\item{xQTL.cred.thres}{The minimum empirical posterior inclusion probability (PIP) used in getting credible sets of xQTL selection. Defaults to 0.95}

\item{xQTL.pip.thres}{If SuSiE fails to find any credible set, the threshold of individual PIP when selecting xQTL. Defaults to 0.5.}

\item{xQTL.pip.min}{The minimum empirical PIP used in purifying variables in each credible set. Defaults to 0.1}

\item{xQTL.Nvec}{The vector of sample sizes of exposures.}

\item{tauvec}{When choosing \code{"IPOD"}, the candidate vector of tuning parameters for the MCP penalty function. Default is \code{seq(3, 30, by=3)}.}

\item{admm.rho}{When choosing \code{"IPOD"}, the tuning parameter in the nested ADMM algorithm. Default is \code{2}.}

\item{ridge.diff}{A ridge.parameter on the differences of causal effect estimate in one credible set. Defaults to \code{10}.}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Defaults to \code{100}.}

\item{max.eps}{Tolerance for stopping criteria. Defaults to \code{0.001}.}

\item{susie.iter}{Number of iterations in SuSiE per iteration. Default is \code{500}.}

\item{ebic.theta}{EBIC factor on causal effect. Default is \code{1}.}

\item{ebic.gamma}{EBIC factor on horizontal pleiotropy Default is \code{2}.}

\item{maxdiff}{The maximum difference between the MRBEE causal estimate and the initial estimator. Defaults to \code{1.5}.}

\item{theta.ini}{Initial value of theta. If \code{FALSE}, the default method is used to estimate. Default is \code{FALSE}.}

\item{gamma.ini}{Initial value of gamma. Default is \code{FALSE}.}

\item{eQTLfitList}{Initial fits of xQTLs of exposures. Default is \code{NULL}.}
}
\value{
A list that contains the results of the MRBEEX with respect to different methods applied:
\describe{
  \item{\code{theta}}{Causal effect estimate.}
  \item{\code{theta.se}}{Standard error of the causal effect estimate.}
  \item{\code{theta.cov}}{Covariance matrix of the causal effect estimate.}
  \item{\code{theta.pip}}{Empirical posterior inclusion probability (PIP) of the causal effect in the subsampling procedure.}
  \item{\code{theta.pratt}}{Pratt index estimate of exposure.}
  \item{\code{gamma}}{Estimate of horizontal pleiotropy.}
  \item{\code{gamma.pratt}}{Pratt index estimate of horizontal pleiotropy.}
  \item{\code{Bic}}{A vector or matrix recording the Bayesian Information Criterion (BIC) values.}
  \item{\code{theta.ini}}{Initial value of \code{theta} used in the estimation procedure.}
  \item{\code{gamma.ini}}{Initial value of \code{gamma} used in the estimation procedure.}
  \item{\code{reliability.adjust}}{Estimated reliability-adjusted values.}
  \item{\code{thetalist}}{List of \code{theta} estimates recorded during each iteration in the subsampling procedure.}
  \item{\code{gammalist}}{List of \code{gamma} estimates recorded during each iteration in the subsampling procedure.}
  \item{var.error}{The variance of residuals.}
  \item{var.error}{The variance of infinitesimal effect.}
}
}
\description{
This function performs multivariable cis-Mendelian randomization that removes weak instrument bias using Bias-corrected Estimating Equations and identifies uncorrelated horizontal pleiotropy (UHP). Additionally, it integrates SuSiE for exposure selection, enhancing interpretability.
}
