% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_snps.R
\name{cluster_snps}
\alias{cluster_snps}
\title{Clustering SNPs based on p-value and proximity with a PLINK C+T file.}
\usage{
cluster_snps(df, window_size = 1e+06)
}
\arguments{
\item{df}{A data.frame containing SNP data with columns for SNP (SNP ID), CHR (chromosome), BP (base pair position), and P (p-value).}

\item{window_size}{An integer specifying the window size around each SNP (in base pairs) within which other SNPs are considered for clustering. Default to 1e6.}
}
\value{
A data.frame containing the clustered SNPs with an additional column 'ClusterSize' indicating the number of SNPs in each cluster.
}
\description{
This function clusters SNPs within a given window size based on their P-value and proximity. It iterates through
each chromosome, finds the SNP with the smallest P-value, and groups all SNPs within the specified window size
around this SNP into a cluster.
}
\details{
The function processes each chromosome independently. It orders the SNPs by their base pair positions, identifies
         the SNP with the smallest P-value, and clusters all SNPs within the specified window size around this SNP. The process
         is repeated until all SNPs are assigned to a cluster.
}
\examples{
df <- data.frame(SNP=c("rs1", "rs2", "rs3", "rs4", "rs5"),
                 CHR=c(1, 1, 1, 1, 2),
                 BP=c(100000, 150000, 200000, 250000, 300000),
                 P=c(0.01, 0.02, 0.03, 0.04, 0.05))
window_size <- 50000
clustered_snps <- cluster_snps(df, window_size)

}
