% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWPT.R
\name{GWPT}
\alias{GWPT}
\title{Genome-Wide Pleiotropy Test}
\usage{
GWPT(by, byse, bX, bXse, Rxy, theta, theta.cov)
}
\arguments{
\item{by}{A vector of effect estimates from the outcome GWAS.}

\item{byse}{A vector of standard errors of effect estimates from the outcome GWAS.}

\item{bX}{A matrix of effect estimates from the exposure GWAS.}

\item{bXse}{A matrix of standard errors of effect estimates from the exposure GWAS.}

\item{Rxy}{The correlation matrix of estimation errors of exposures and outcome GWAS. The last column corresponds to the outcome.}

\item{theta}{The causal effect estimate.}

\item{theta.cov}{The covariance matrix of the causal effect estimate.}
}
\value{
A list with two components:
\item{BETA}{The estimated residual values.}
\item{SE}{The standard errors of the residual estimates.}
}
\description{
This function performs a genome-wide pleiotropy test (GWPT) after Mendelian randomization. It offers an option for a two-mixture model, where the residual is chosen as the smaller one resulting from the two causal effect estimates from two mixtures.
}
