% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWPT_Mixture.R
\name{GWPT_Mixture}
\alias{GWPT_Mixture}
\title{Genome-Wide Pleiotropy Test for mixture model}
\usage{
GWPT_Mixture(
  by,
  byse,
  bX,
  bXse,
  Rxy,
  theta1,
  theta.cov1,
  theta2,
  theta.cov2,
  LD.block
)
}
\arguments{
\item{by}{A vector of effect estimates from the outcome GWAS.}

\item{byse}{A vector of standard errors of effect estimates from the outcome GWAS.}

\item{bX}{A matrix of effect estimates from the exposure GWAS.}

\item{bXse}{A matrix of standard errors of effect estimates from the exposure GWAS.}

\item{Rxy}{The correlation matrix of estimation errors of exposures and outcome GWAS. The last column corresponds to the outcome.}

\item{theta1}{The causal effect estimate of the first mixture.}

\item{theta.cov1}{The covariance matrix of the causal effect estimate of the first mixture.}

\item{theta2}{The causal effect estimate of the second mixture.}

\item{theta.cov2}{The covariance matrix of the causal effect estimate of the second mixture.}

\item{LD.block}{A vector of indices of LD blocks.}
}
\value{
A list with two components:
\item{BETA}{The estimated residual values.}
\item{SE}{The standard errors of the residual estimates.}
}
\description{
This function performs a genome-wide pleiotropy test (GWPT) after Mendelian randomization. It offers an option for a two-mixture model, where the residual is chosen as the smaller one resulting from the two causal effect estimates from two mixtures.
}
