% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBEE_SuSiE.R
\name{MRBEE_SuSiE}
\alias{MRBEE_SuSiE}
\title{Mendelian randomization with bias-correction estimating equation: selecting horizontal pleiotropy via IPOD and selecting exposures via SuSiE.}
\usage{
MRBEE_SuSiE(
  by,
  bX,
  byse,
  bXse,
  LD = "identity",
  Rxy,
  cluster.index = c(1:length(by)),
  Nmin = F,
  Lvec = c(1:min(10, nrow(bX))),
  pip.thres = 0.5,
  tauvec = seq(3, 50, by = 2),
  max.iter = 100,
  max.eps = 0.001,
  susie.iter = 100,
  ebic.theta = 0,
  ebic.gamma = 1,
  empirical.variance.lower = 0.2,
  empirical.variance.upper = 100,
  reliability.thres = 0.8,
  rho = 2,
  maxdiff = 1.5,
  sampling.time = 100,
  sampling.frac = 0.5,
  sampling.iter = 5,
  theta.ini = F,
  gamma.ini = F
)
}
\arguments{
\item{by}{A vector (n x 1) of GWAS effect sizes for the outcome.}

\item{bX}{A matrix (n x p) of GWAS effect sizes for p exposures.}

\item{byse}{A vector (n x 1) of standard errors for the GWAS effect sizes of the outcome.}

\item{bXse}{A matrix (n x p) of standard errors for the GWAS effect sizes of the exposures.}

\item{LD}{A matrix representing the linkage disequilibrium (LD) among instrumental variables.}

\item{Rxy}{A matrix (p+1 x p+1) of the correlation matrix including p exposures and the outcome. Outcome should be the last column.}

\item{cluster.index}{A vector indicating the cluster membership for each instrumental variable. This is used in standard error estimation.}

\item{Nmin}{Optional; the minimum sample size for the GWAS if not provided, defaults to the number of instrumental variables.}

\item{Lvec}{The number of single effects used in SuSiE, defaults to c(1:min(5,nrow(bX))).}

\item{pip.thres}{The threshold of PIP in SuSiE, below which the causal effect estimate will be set to zero, defaults to 0.5.}

\item{tauvec}{A vector of tuning parameters for penalizing horizontal pleiotropy in the IPOD algorithm.}

\item{max.iter}{The maximum number of iterations allowed for convergence of the causal effect estimates.}

\item{max.eps}{The tolerance level for convergence; iteration stops when changes are below this threshold.}

\item{susie.iter}{The maximum number of iterations allowed for convergence of SuSiE program, defaults to 50.}

\item{ebic.theta}{The penalty factor for extended Bayesian Information Criterion (eBIC) adjustments on causal effects}

\item{ebic.gamma}{The penalty factor for eBIC adjustments on pleiotropy.}

\item{empirical.variance.lower}{The lower boundary of empirical variance estimate of residual, defaults to 0.5.}

\item{empirical.variance.upper}{The upper boundary of empirical variance estimate of residual, defaults to 2.}

\item{reliability.thres}{A threshold on bias-correction term, defaults to 0.8.}

\item{rho}{The penalty multiplier used in the ADMM algorithm within the IPOD framework.}

\item{maxdiff}{The maximum allowed difference ratio between iterative causal estimates and initial estimations for stabilization.}

\item{sampling.time}{The number of subsampling iterations used to estimate the standard error of the causal effect estimate. Defaults to 100. When set to 0, a sandwich formula is applied for the estimation.}

\item{sampling.frac}{The fraction of the data to be used in each subsampling iteration. Defaults to 0.5, meaning that 50\% of the data is used in each iteration.}

\item{sampling.iter}{The number of iteration of MRBEE_SuSiE to be used in each subsampling iteration. Defaults to 5.}

\item{theta.ini}{Initial estimates for the causal effects; defaults to FALSE, indicating automatic initialization.}

\item{gamma.ini}{Initial estimates for horizontal pleiotropy effects; also defaults to FALSE for automatic setup.}
}
\value{
A list containing detailed results of the analysis, including estimated causal effects, pleiotropy effects, their respective standard errors, and Bayesian Information Criterion (BIC) scores, among other metrics.
}
\description{
Detailed description of the function goes here.
}
