% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBEEX_UV.R
\name{MRBEEX_UV}
\alias{MRBEEX_UV}
\title{Univariable Mendelian Randomization using Bias-corrected Estimating Equations}
\usage{
MRBEEX_UV(
  by,
  bX,
  byse,
  bXse,
  LD = "identity",
  Rxy,
  cluster.index = c(1:length(by)),
  reliability.thres = 0.6,
  Method = "IPOD",
  tauvec = seq(4, 8, by = 2),
  admm.rho = 2,
  ebic.gamma = 1,
  Kvec = seq(1, length(bX)/2, by = 2),
  max.iter = 100,
  max.eps = 0.001,
  maxdiff = 3,
  sampling.time = 1000,
  sampling.iter = 20,
  theta.ini = F,
  gamma.ini = F
)
}
\arguments{
\item{by}{A vector of effect estimates from the outcome GWAS.}

\item{bX}{A vector of effect estimates from the exposure GWAS.}

\item{byse}{A vector of standard errors of effect estimates from the outcome GWAS.}

\item{bXse}{A vector of standard errors of effect estimates from the exposure GWAS.}

\item{LD}{The linkage disequilibrium (LD) matrix. Default is the identity matrix, assuming independent instrumental variables (IVs).}

\item{Rxy}{The correlation matrix of estimation errors of exposures and outcome GWAS. The last element corresponds to the outcome.}

\item{cluster.index}{A vector indicating the LD block indices each IV belongs to. The length is equal to the number of IVs, and values are the LD block indices.}

\item{reliability.thres}{A threshold for the minimum value of the reliability ratio. If the original reliability ratio is less than this threshold, only part of the estimation error is removed so that the working reliability ratio equals this threshold. Default is \code{0.6}.}

\item{Method}{Method for handling horizontal pleiotropy. Options are \code{"IPOD"} and \code{"Greedy"}.}

\item{tauvec}{When choosing \code{"IPOD"}, the candidate vector of tuning parameters for the MCP penalty function. Default is \code{seq(4, 8, by=2)}.}

\item{admm.rho}{When choosing \code{"IPOD"}, the tuning parameter in the nested ADMM algorithm. Default is \code{2}.}

\item{ebic.gamma}{EBIC factor on horizontal pleiotropy Default is \code{1}.}

\item{Kvec}{When choosing \code{"Greedy"}, the candidate vector of number of pleiotropy in the greedy search algorithm. Default is \code{seq(1, length(bX)/2, by=2)}.}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Defaults to \code{100}.}

\item{max.eps}{Tolerance for stopping criteria. Defaults to \code{0.001}.}

\item{maxdiff}{The maximum difference between the MRBEE causal estimate and the initial estimator. Defaults to \code{3}.}

\item{sampling.time}{Number of resampling times. Default is \code{1000}.}

\item{sampling.iter}{Number of iterations per resampling. Default is \code{20}.}

\item{theta.ini}{Initial value of theta. If \code{FALSE}, the default method is used to estimate. Default is \code{FALSE}.}

\item{gamma.ini}{Initial value of gamma. Default is \code{FALSE}.}
}
\value{
A list containing the results of the MRBEEX.UV analysis:
\describe{
  \item{\code{theta}}{Causal effect estimate.}
  \item{\code{theta.se}}{Standard error of the causal effect estimate.}
  \item{\code{theta.cov}}{Covariance matrix of the causal effect estimate.}
  \item{\code{theta.bootstrap}}{Resampled causal effect estimates.}
  \item{\code{gamma}}{Estimate of horizontal pleiotropy.}
  \item{\code{Bic}}{A vector or matrix recording the Bayesian Information Criterion (BIC) values.}
  \item{\code{theta.ini}}{Initial value of \code{theta} used in the estimation procedure.}
  \item{\code{gamma.ini}}{Initial value of \code{gamma} used in the estimation procedure.}
  \item{\code{reliability.adjust}}{Estimated reliability-adjusted values.}
  \item{\code{thetalist}}{List of \code{theta} estimates recorded during each iteration in the resampling procedure.}
  \item{\code{gammalist}}{List of \code{gamma} estimates recorded during each iteration in the resampling procedure.}
}
}
\description{
This function removes weak instrument bias using Bias-corrected Estimating Equations and identifies uncorrelated horizontal pleiotropy (UHP) and correlated horizontal pleiotropy (CHP) through IPOD or greedy search.
}
