% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_Prediction.R
\name{Sparse_Prediction}
\alias{Sparse_Prediction}
\title{Sparse Prediction of xQTL Effects using SuSiE}
\usage{
Sparse_Prediction(
  bX,
  bXse,
  LD,
  xQTL.Nvec,
  xQTL.max.L = 10,
  xQTL.weight = NULL,
  xQTL.cred.thres = 0.95,
  estimate_residual_method = "MoM",
  unmappable_effects = "inf"
)
}
\arguments{
\item{bX}{A matrix of effect estimates from the exposure GWAS.}

\item{bXse}{A matrix of standard errors of effect estimates from the exposure GWAS.}

\item{LD}{The linkage disequilibrium (LD) matrix.}

\item{xQTL.Nvec}{The vector of sample sizes of exposures.}

\item{xQTL.max.L}{The maximum number of single effects used to estimate the xQTL effects. Defaults to 10.}

\item{xQTL.weight}{The vector of weights used in specifying the prior weights of SuSiE. Defaults to \code{NULL}.}

\item{xQTL.cred.thres}{The minimum empirical posterior inclusion probability (PIP) used in getting credible sets of xQTL selection. Defaults to \code{0.95}.}

\item{estimate_residual_method}{The method used for estimating residual variance. For the original SuSiE model, "MLE" and "MoM" estimation is equivalent, but for the infinitesimal model, "MoM" is more stable. We recommend using "Servin_Stephens" when n < 80 for improved coverage.}

\item{unmappable_effects}{The method for modeling unmappable effects: "none", "inf".}
}
\value{
A list containing the results of the MRBEEX analysis using different methods:
\describe{
  \item{\code{bXest}}{A matrix where each column represents \code{R * beta_PLS} for each exposure.}
  \item{\code{bXestse}}{A matrix where each column contains the standard errors of \code{R * beta_PLS} for each exposure.}
  \item{\code{bXest0}}{A matrix where each column represents \code{beta_PLS} for each exposure.}
  \item{\code{bXestse0}}{A matrix where each column contains the standard errors of \code{beta_PLS} for each exposure.}
  \item{\code{xQTLfitList}}{A list containing the xQTL selection results, which can be used in \code{CisMRBEEX} by setting \code{xQTLfitList = xQTLfitList}.}
}
}
\description{
This function performs informative xQTL selection and sparse prediction using SuSiE.
}
