% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary.R
\name{summary_generation}
\alias{summary_generation}
\title{Generating simulated data for Mendelian randomization simulation}
\usage{
summary_generation(
  theta,
  m,
  Rbb,
  Ruv,
  Rnn,
  Nxy,
  Hxy,
  LD = "identity",
  non.zero.frac,
  UHP.frac = 0,
  CHP.frac = 0,
  UHP.var = 0.5,
  UHP.dis = "ash",
  CHP.effect = c(1, rep(0, length(theta) - 1)),
  effect.dis = "normal",
  cluster.index,
  prop = c(0.9, 0.09, 0.009, 9e-04),
  var = c(0.35, 0.4, 0.2, 0.05)
)
}
\arguments{
\item{theta}{An (px1) vector of causal effects.}

\item{m}{The number of instrumental variables (IVs).}

\item{Rbb}{An (pxp) correlation matrix of genetic effects.}

\item{Ruv}{An ((p+1)x(p+1)) correlation matrix of residuals in outcome and exposures; the outcome is the last one.}

\item{Rnn}{An ((p+1)x(p+1)) correlation matrix of sample overlap; the outcome is the last one.}

\item{Nxy}{An ((p+1)x1) vector of GWAS sample sizes; the sample size of the outcome is the last one.}

\item{Hxy}{An ((p+1)x1) vector of heritabilities; the outcome is the last one.}

\item{LD}{An (mxm) correlation matrix of the IVs or "identity" indicating independent IVs.}

\item{non.zero.frac}{An (px1) vector with all entries in (0,1]; each entry is the probability of deltaj such that betaj=betaj'*deltaj.}

\item{UHP.frac}{A number indicating the fraction of IVs affected by UHP.}

\item{CHP.frac}{A number indicating the fraction of IVs affected by CHP.}

\item{UHP.var}{A number indicating the variance attributed to UHP.}

\item{UHP.dis}{Distribution of pleiotropy effects: "ash" (default), "normal", "uniform",  "t" distribution (with degree of freedom 5).}

\item{CHP.effect}{A vector of effects corresponding to the variables correlated with the correlated horizontal pleiotropy.}

\item{effect.dis}{Distribution of genetic effects: "normal" (default), "uniform",  "t" distribution (with degree of freedom 5).}

\item{cluster.index}{The indices of LD block.}

\item{prop}{When UHP.dis="ash", a numeric vector of mixing proportions (weights). These are normalized. Default: c(0.9, 0.09, 0.009, 0.0009).}

\item{var}{When UHP.dis="ash", a numeric vector where var[i] represents the variance contribution of component i: V_i = prop[i] * sigma_i^2. Default: c(0.35, 0.4, 0.2, 0.05).}
}
\value{
A list containing simulated GWAS effect sizes for exposures (bX), their standard errors (bXse),
        the GWAS effect size for the outcome (by), its standard error (byse), the pleiotropy effects (pleiotropy), and the true effects.
}
\description{
This function generates simulated data for Mendelian Randomization (MR) analysis,
considering genetic effects, estimation errors, and horizontal pleiotropy.
It allows for different distributions of genetic effects and pleiotropy,
and accommodates both independent and correlated instrumental variables (IVs).
}
