% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CisMRBEEX.R
\name{CisMRBEEX}
\alias{CisMRBEEX}
\title{Cis Multivariable Mendelian Randomization using Bias-corrected Estimating Equations}
\usage{
CisMRBEEX(
  by,
  bX,
  byse,
  bXse,
  LD,
  Rxy,
  model.infinitesimal = F,
  reliability.thres = 0.75,
  Lvec = c(1:5),
  causal.pip.thres = 0.2,
  xQTL.method = "SuSiE",
  xQTL.selection.rule = "top_K",
  top_K = 1,
  xQTL.pip.min = 0.2,
  xQTL.max.L = 10,
  xQTL.cred.thres = 0.95,
  xQTL.pip.thres = 0.5,
  xQTL.Nvec,
  tauvec = seq(3, 30, by = 3),
  xQTL.weight = NULL,
  outlier.switch = T,
  Annotation = NULL,
  output.labels = NULL,
  carma.iter = 5,
  carma.inner.iter = 5,
  xQTL.max.num = 10,
  carma.epsilon.threshold = 0.001,
  admm.rho = 2,
  ridge.diff = 1000,
  max.iter = 100,
  max.eps = 0.001,
  susie.iter = 500,
  coverage.xQTL = 0.95,
  coverage.causal = 0.95,
  ebic.theta = 0,
  ebic.gamma = 1,
  theta.ini = F,
  gamma.ini = F,
  xQTLfitList = NULL,
  verbose = T
)
}
\arguments{
\item{by}{A vector of effect estimates from the outcome GWAS.}

\item{bX}{A matrix of effect estimates from the exposure GWAS.}

\item{byse}{A vector of standard errors of effect estimates from the outcome GWAS.}

\item{bXse}{A matrix of standard errors of effect estimates from the exposure GWAS.}

\item{LD}{The linkage disequilibrium (LD) matrix.}

\item{Rxy}{The correlation matrix of estimation errors of exposures and outcome GWAS. The last column corresponds to the outcome.}

\item{model.infinitesimal}{An indicator of whether using REML to model infinitesimal effects. Defaults to \code{F}.}

\item{reliability.thres}{A threshold for the minimum value of the reliability ratio. If the original reliability ratio is less than this threshold, only part of the estimation error is removed so that the working reliability ratio equals this threshold.}

\item{Lvec}{When SuSiE is used, the candidate vector for the number of single effects. Default is \code{c(1:min(10, nrow(bX)))}.}

\item{causal.pip.thres}{A threshold of minimum posterior inclusion probability. Default is \code{0.2}.}

\item{xQTL.method}{The method used in purifying the xQTLs. SuSiE or CARMA can be used here, where the latter can be more accurate but much most computationally costly. Defaults is SuSiE.}

\item{xQTL.selection.rule}{The method for purifying informative xQTLs within each credible set. Options include "minimum_pip", which selects all variables with PIPs exceeding a specified threshold, and "top_K", which ensures at least K variables are selected based on their PIP ranking. Defaults to "top_K".}

\item{top_K}{The maximum number of variables selected in each credible sets. Defaults to 1.}

\item{xQTL.pip.min}{The minimum empirical PIP used in purifying variables in each credible set. Defaults to \code{0.2}.}

\item{xQTL.max.L}{When choosing \code{"SuSiE"}, the maximum number of L in estimating the xQTL effects. Defaults to 10.}

\item{xQTL.cred.thres}{When choosing \code{"SuSiE"}, the minimum empirical posterior inclusion probability (PIP) used in getting credible sets of xQTL selection. Defaults to \code{0.95}.}

\item{xQTL.pip.thres}{The threshold of individual PIP when selecting xQTL. Defaults to \code{0.5}.}

\item{xQTL.Nvec}{When choosing \code{"SuSiE"}, the vector of sample sizes of exposures.}

\item{tauvec}{The candidate vector of tuning parameters for the MCP penalty function. Default is \code{seq(3, 30, by=3)}.}

\item{xQTL.weight}{When choosing \code{"SuSiE"}, the vector of weights used in specifying the prior weights of SuSiE. Defaults to \code{NULL}.}

\item{outlier.switch}{When choosing \code{"CARMA"}, an indicator of whether turning on outlier detection. Defaults to \code{F}.}

\item{Annotation}{When choosing \code{"CARMA"}, the annotation matrix of SNP. Default is NULL.}

\item{output.labels}{When choosing \code{"CARMA"}, output directory where output will be written while CARMA is running. Defaults to \code{NULL}, meaning that a temporary folder will be created and automatically deleted upon completion of the computation.}

\item{carma.iter}{When choosing \code{"CARMA"}, the maximum iterations for EM algorithm to run. Defaults to 5.}

\item{carma.inner.iter}{When choosing \code{"CARMA"}, the maximum iterations for Shotgun algorithm to run per iteration within EM algorithm. Defaults to 5.}

\item{xQTL.max.num}{When choosing \code{"CARMA"}, the maximum number of causal variants assumed per locus, which is similar to the number of single effects in SuSiE. Defaults to 10.}

\item{carma.epsilon.threshold}{When choosing \code{"CARMA"}, the convergence threshold measured by average of Bayes factors. Defaults to \code{1e-3}.}

\item{admm.rho}{The tuning parameter in the nested ADMM algorithm. Default is \code{2}.}

\item{ridge.diff}{A ridge.parameter on the differences of causal effect estimate in one credible set. Defaults to \code{10}.}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Defaults to \code{100}.}

\item{max.eps}{Tolerance for stopping criteria. Defaults to \code{0.001}.}

\item{susie.iter}{Number of iterations in SuSiE per iteration. Default is \code{500}.}

\item{coverage.xQTL}{The coverage of defining a credible set in xQTL selection. Defaults to \code{0.95}.}

\item{coverage.causal}{The coverage of defining a credible set in cis-MRBEE. Defaults to \code{0.95}.}

\item{ebic.theta}{EBIC factor on causal effect. Default is \code{1}.}

\item{ebic.gamma}{EBIC factor on horizontal pleiotropy Default is \code{2}.}

\item{theta.ini}{Initial value of theta. If \code{FALSE}, the default method is used to estimate. Default is \code{FALSE}.}

\item{gamma.ini}{Initial value of gamma. Default is \code{FALSE}.}

\item{xQTLfitList}{Initial fits of xQTLs for exposures. This should be a list. Each component corresponds to the susie.fit of each exposure when xQTL.method = "SuSiE". When xQTL.method = "CARMA", this should be the list of results from a CARMA analysis. Users can customize additional SuSiE or CARMA parameters to improve performance. Default is \code{NULL}.}

\item{verbose}{A logical indicator of whether to display the execution time of the method. Default is \code{T}.}
}
\value{
A list that contains the results of the MRBEEX with respect to different methods applied:
\describe{
  \item{\code{theta}}{Causal effect estimate.}
  \item{\code{theta.se}}{Standard error of the causal effect estimate.}
  \item{\code{theta.cov}}{Covariance matrix of the causal effect estimate.}
  \item{\code{theta.pip}}{Empirical posterior inclusion probability (PIP) of the causal effect in the subsampling procedure.}
  \item{\code{theta.pratt}}{Pratt index estimate of exposure.}
  \item{\code{susie.theta}}{The fit of causal effect resulted from susie.}
  \item{\code{gamma}}{Estimate of horizontal pleiotropy.}
  \item{\code{gamma.pratt}}{Pratt index estimate of horizontal pleiotropy.}
  \item{\code{Bic}}{A vector or matrix recording the Bayesian Information Criterion (BIC) values.}
  \item{\code{theta.ini}}{Initial value of \code{theta} used in the estimation procedure.}
  \item{\code{gamma.ini}}{Initial value of \code{gamma} used in the estimation procedure.}
  \item{\code{reliability.adjust}}{Estimated reliability-adjusted values.}
  \item{\code{thetalist}}{List of \code{theta} estimates recorded during each iteration in the subsampling procedure.}
  \item{\code{gammalist}}{List of \code{gamma} estimates recorded during each iteration in the subsampling procedure.}
  \item{var.error}{The variance of residuals.}
  \item{var.error}{The variance of infinitesimal effect.}
  \item{estimated.reliability.ratio}{The estimated reliability ratios of exposures.}
  \item{xQTLfitList}{The results of sparse predictions of exposures, yielded by SuSiE or CARMA.}
}
}
\description{
This function performs multivariable cis-Mendelian randomization that removes weak instrument bias using Bias-corrected Estimating Equations and identifies uncorrelated horizontal pleiotropy (UHP). Additionally, it integrates SuSiE for exposure selection, enhancing interpretability.
}
