% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBEE_TL.R
\name{MRBEE_TL}
\alias{MRBEE_TL}
\title{Estimate Non-Transferable Causal Effect with MRBEE and SuSiE}
\usage{
MRBEE_TL(
  by,
  bX,
  byse,
  bXse,
  Rxy,
  LD = "identity",
  cluster.index = c(1:length(by)),
  theta.source,
  theta.source.cov,
  tauvec = seq(3, 30, 3),
  Lvec = c(1:6),
  admm.rho = 3,
  ebic.delta = 1,
  ebic.gamma = 2,
  transfer.coef = 1,
  susie.iter = 200,
  pip.thres = 0.3,
  max.iter = 50,
  max.eps = 1e-04,
  reliability.thres = 0.8,
  ridge.diff = 100,
  sampling.time = 100,
  sampling.iter = 10
)
}
\arguments{
\item{by}{A vector (n x 1) of the GWAS effect size of outcome.}

\item{bX}{A matrix (n x p) of the GWAS effect sizes of p exposures.}

\item{byse}{A vector (n x 1) of the GWAS effect size SE of outcome.}

\item{bXse}{A matrix (n x p) of the GWAS effect size SEs of p exposures.}

\item{Rxy}{A matrix (p+1 x p+1) of the correlation matrix of the p exposures and outcome. The first one should be the transferred linear predictor and last one should be the outcome.}

\item{LD}{The linkage disequilibrium (LD) matrix. Default is the identity matrix, assuming independent instrumental variables (IVs).}

\item{cluster.index}{A vector indicating the LD block indices each IV belongs to. The length is equal to the number of IVs, and values are the LD block indices.}

\item{theta.source}{A vector (p x 1) of the causal effect estimate learning from the source data.}

\item{theta.source.cov}{A matrix (p x p) of the covariance matrix of the causal effect estimate learning from the source data.}

\item{tauvec}{The candidate vector of tuning parameters for the MCP penalty function. Default is \code{seq(3, 30, by=3)}.}

\item{Lvec}{A vector of the number of single effects used in SuSiE. Default is \code{c(1:6)}.}

\item{admm.rho}{When choosing \code{"IPOD"}, the tuning parameter in the nested ADMM algorithm. Default is \code{2}.}

\item{ebic.delta}{A scale of tuning parameter of causal effect estimate in extended BIC. Default is \code{1}.}

\item{ebic.gamma}{A scale of tuning parameter of horizontal pleiotropy in extended BIC. Default is \code{2}.}

\item{transfer.coef}{A scale of transfer.coef of theta.source to theta.target. Default is \code{1}.}

\item{susie.iter}{A scale of the maximum number of iterations used in SuSiE. Default is \code{200}.}

\item{pip.thres}{A scale of PIP theshold for calibyating causality used in SuSiE. Default is \code{0.3}.}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Default is \code{50}.}

\item{max.eps}{Tolerance for stopping criteria. Default is \code{1e-4}.}

\item{reliability.thres}{A scale of threshold for the minimum value of the reliability ratio. If the original reliability ratio is less than this threshold, only part of the estimation error is removed so that the working reliability ratio equals this threshold. Default is \code{0.8}.}

\item{ridge.diff}{A scale of parameter on the differences of causal effect estimate in one credible set. Defaults to \code{10}.}

\item{sampling.time}{A scale of number of subsampling in estimating the standard error. Default is \code{100}.}

\item{sampling.iter}{A scale of iteration in subsampling in estimating the standard error. Default is \code{10}.}
}
\value{
A list containing the estimated causal effect, its covariance, and pleiotropy.
}
\description{
This function estimates the non-transferable causal effect using a bias-correction estimating equation, considering potential pleiotropy and measurement errors, and using SuSiE to select the non-transferable causal effect.
}
