% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRBEE_TL.R
\name{MRBEE_TL}
\alias{MRBEE_TL}
\title{Estimate Non-Transferable Causal Effect with MRBEE and SuSiE}
\usage{
MRBEE_TL(
  by,
  bX,
  byse,
  bXse,
  Rxy,
  LD = "identity",
  cluster.index = c(1:length(by)),
  group.penalize = F,
  group.index = NULL,
  group.diff = 100,
  theta.source,
  theta.source.cov,
  tauvec = seq(4, 8, 0.5),
  Lvec = c(1:6),
  standardize = T,
  admm.rho = 2,
  ebic.delta = 0,
  ebic.gamma = 1,
  transfer.coef = 1,
  susie.iter = 200,
  pip.thres = 0.25,
  pip.min = 0.1,
  cred.pip.thres = 0.95,
  max.iter = 50,
  coverage.causal = 0.95,
  max.eps = 1e-06,
  reliability.thres = 0.5,
  ridge.diff = 100,
  prob_shrinkage_coef = 0.5,
  prob_shrinkage_size = 4,
  estimate_residual_method = "MoM",
  sampling.strategy = "bootstrap",
  sampling.time = 300,
  sampling.iter = 25,
  ldsc = NULL,
  gcov = NULL
)
}
\arguments{
\item{by}{A vector (n x 1) of the GWAS effect size of outcome.}

\item{bX}{A matrix (n x p) of the GWAS effect sizes of p exposures.}

\item{byse}{A vector (n x 1) of the GWAS effect size SE of outcome.}

\item{bXse}{A matrix (n x p) of the GWAS effect size SEs of p exposures.}

\item{Rxy}{A matrix (p+1 x p+1) of the correlation matrix of the p exposures and outcome. The first one should be the transferred linear predictor and last one should be the outcome.}

\item{LD}{The linkage disequilibrium (LD) matrix. Default is the identity matrix, assuming independent instrumental variables (IVs).}

\item{cluster.index}{A vector indicating the LD block indices each IV belongs to. The length is equal to the number of IVs, and values are the LD block indices.}

\item{group.penalize}{An indicator of whether using difference penalty to penalize highly correlated exposures. Defaults to \code{F}.}

\item{group.index}{A vector of the group index of exposure. Defaults to \code{NULL}.}

\item{group.diff}{The tuning penalizing difference of highly correlated exposure prediction. Defaults to \code{10}.}

\item{theta.source}{A vector (p x 1) of the causal effect estimate learning from the source data.}

\item{theta.source.cov}{A matrix (p x p) of the covariance matrix of the causal effect estimate learning from the source data.}

\item{tauvec}{The candidate vector of tuning parameters for the MCP penalty function. Default is \code{seq(4, 8, by=0.5)}.}

\item{Lvec}{A vector of the number of single effects used in SuSiE. Default is \code{c(1:6)}.}

\item{standardize}{If standardize = TRUE, standardize the columns of X to unit variance prior to fitting (or equivalently standardize XtX and Xty to have the same effect) in SuSiE. Note that scaled_prior_variance specifies the prior on the coefficients of X after standardization (if it is performed). If you do not standardize, you may need to think more carefully about specifying scaled_prior_variance. Whatever your choice, the coefficients returned by coef are given for X on the original input scale. Any column of X that has zero variance is not standardized.}

\item{admm.rho}{When choosing \code{"IPOD"}, the tuning parameter in the nested ADMM algorithm. Default is \code{2}.}

\item{ebic.delta}{A scale of tuning parameter of causal effect estimate in extended BIC. Default is \code{0}.}

\item{ebic.gamma}{A scale of tuning parameter of horizontal pleiotropy in extended BIC. Default is \code{1}.}

\item{transfer.coef}{A scale of transfer.coef of theta.source to theta.target. Default is \code{1}.}

\item{susie.iter}{A scale of the maximum number of iterations used in SuSiE. Default is \code{200}.}

\item{pip.thres}{Posterior inclusion probability (PIP) threshold. Individual PIPs less than this value will be shrunk to zero. Default is \code{0.25}.}

\item{pip.min}{The minimum empirical PIP used in purifying variables in each credible set. Defaults to \code{0.1}.}

\item{cred.pip.thres}{The threshold of PIP of each credible set. Defaults to \code{0.95}.}

\item{max.iter}{Maximum number of iterations for causal effect estimation. Default is \code{50}.}

\item{coverage.causal}{The coverage of defining a credible set in MRBEEX when \code{use.susie = T}. Defaults to \code{0.95}.}

\item{max.eps}{Tolerance for stopping criteria. Default is \code{1e-4}.}

\item{reliability.thres}{A scale of threshold for the minimum value of the reliability ratio. If the original reliability ratio is less than this threshold, only part of the estimation error is removed so that the working reliability ratio equals this threshold. Default is \code{0.6}.}

\item{ridge.diff}{A scale of parameter on the differences of causal effect estimate in one credible set. Defaults to \code{10}.}

\item{prob_shrinkage_coef}{Shrinkage coefficient (alpha) used to smooth block sampling probabilities within each group. 0 = no shrinkage; 1 = full shrinkage to group median.}

\item{prob_shrinkage_size}{Number of blocks per smoothing group (e.g., 3–5). Blocks are sorted by weight and grouped before applying shrinkage.}

\item{estimate_residual_method}{The method used for estimating residual variance. For the original SuSiE model, "MLE" and "MoM" estimation is equivalent, but for the infinitesimal model, "MoM" is more stable.}

\item{sampling.strategy}{"bootstrap" or "subsampling" (0.5 sample without replacement).}

\item{sampling.time}{A scale of number of subsampling in estimating the standard error. Default is \code{100}.}

\item{sampling.iter}{A scale of iteration in subsampling in estimating the standard error. Default is \code{10}.}

\item{ldsc}{A vector (n x 1) of the LDSCs of the IVs.}

\item{gcov}{A matrix (p+1 x p+1) of the per-snp genetic covariance matrix of the p exposures and outcome. The last one should be the outcome.}
}
\value{
A list containing the estimated causal effect, its covariance, and pleiotropy.
}
\description{
This function estimates the non-transferable causal effect using a bias-correction estimating equation, considering potential pleiotropy and measurement errors, and using SuSiE to select the non-transferable causal effect.
}
