% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_Prediction.R
\name{Sparse_Prediction}
\alias{Sparse_Prediction}
\title{Sparse Prediction of xQTL Effects using SuSiE and CARMA}
\usage{
Sparse_Prediction(
  bX,
  bXse,
  LD,
  xQTL.method = "SuSiE",
  xQTL.selection.rule = "top_K",
  top_K = 1,
  xQTL.pip.min = 0.2,
  ridge.diff = 100,
  xQTL.max.L = 10,
  xQTL.cred.thres = 0.95,
  xQTL.pip.thres = 0.5,
  xQTL.Nvec,
  xQTL.weight = NULL,
  outlier.switch = T,
  Annotation = NULL,
  output.labels = NULL,
  carma.iter = 5,
  carma.inner.iter = 5,
  xQTL.max.num = 10,
  carma.epsilon.threshold = 0.001
)
}
\arguments{
\item{bX}{A matrix of effect estimates from the exposure GWAS.}

\item{bXse}{A matrix of standard errors of effect estimates from the exposure GWAS.}

\item{LD}{The linkage disequilibrium (LD) matrix.}

\item{xQTL.method}{The method used in purifying the xQTLs. SuSiE or CARMA can be used here, where the latter can be more accurate but much most computationally costly. Defaults is SuSiE.}

\item{xQTL.selection.rule}{The method for purifying informative xQTLs within each credible set. Options include "minimum_pip", which selects all variables with PIPs exceeding a specified threshold, and "top_K", which ensures at least K variables are selected based on their PIP ranking. Defaults to "top_K".}

\item{top_K}{The maximum number of variables selected in each credible sets. Defaults to 1.}

\item{xQTL.pip.min}{The minimum empirical PIP used in purifying variables in each credible set. Defaults to \code{0.2}.}

\item{ridge.diff}{A ridge.parameter on the differences of causal effect estimate in one credible set. Defaults to \code{100}.}

\item{xQTL.max.L}{When choosing \code{"SuSiE"}, the maximum number of L in estimating the xQTL effects. Defaults to 10.}

\item{xQTL.cred.thres}{When choosing \code{"SuSiE"}, the minimum empirical posterior inclusion probability (PIP) used in getting credible sets of xQTL selection. Defaults to \code{0.95}.}

\item{xQTL.pip.thres}{When choosing \code{"SuSiE"}, the threshold of individual PIP when selecting xQTL. Defaults to \code{0.5}.}

\item{xQTL.Nvec}{When choosing \code{"SuSiE"}, the vector of sample sizes of exposures.}

\item{xQTL.weight}{When choosing \code{"SuSiE"}, the vector of weights used in specifying the prior weights of SuSiE. Defaults to \code{NULL}.}

\item{outlier.switch}{When choosing \code{"CARMA"}, an indicator of whether turning on outlier detection. Defaults to \code{F}.}

\item{Annotation}{When choosing \code{"CARMA"}, the annotation matrix of SNP. Default is NULL.}

\item{output.labels}{When choosing \code{"CARMA"}, output directory where output will be written while CARMA is running. Defaults to \code{NULL}, meaning that a temporary folder will be created and automatically deleted upon completion of the computation.}

\item{carma.iter}{When choosing \code{"CARMA"}, the maximum iterations for EM algorithm to run. Defaults to 5.}

\item{carma.inner.iter}{When choosing \code{"CARMA"}, the maximum iterations for Shotgun algorithm to run per iteration within EM algorithm. Defaults to 5.}

\item{xQTL.max.num}{When choosing \code{"CARMA"}, the maximum number of causal variants assumed per locus, which is similar to the number of single effects in SuSiE. Defaults to 10.}

\item{carma.epsilon.threshold}{When choosing \code{"CARMA"}, the convergence threshold measured by average of Bayes factors. Defaults to \code{1e-3}.}
}
\value{
A list containing the results of the MRBEEX analysis using different methods:
\describe{
  \item{\code{bXest}}{A matrix where each column represents \code{R * beta_PLS} for each exposure.}
  \item{\code{bXestse}}{A matrix where each column contains the standard errors of \code{R * beta_PLS} for each exposure.}
  \item{\code{bXest0}}{A matrix where each column represents \code{beta_PLS} for each exposure.}
  \item{\code{bXestse0}}{A matrix where each column contains the standard errors of \code{beta_PLS} for each exposure.}
  \item{\code{xQTLfitList}}{A list containing the xQTL selection results, which can be used in \code{CisMRBEEX} by setting \code{xQTLfitList = xQTLfitList}.}
}
}
\description{
This function performs infomative xQTL selection and sparse prediction using SuSiE and CARMA.
}
