% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparse_Prediction.R
\name{Sparse_Prediction}
\alias{Sparse_Prediction}
\title{Sparse Prediction of xQTL Effects using SuSiE and CARMA}
\usage{
Sparse_Prediction(
  bX,
  bXse,
  LD,
  xQTL.Nvec,
  xQTL.method = "SuSiE",
  xQTL.max.L = 10,
  xQTL.weight = NULL,
  xQTL.cred.thres = 0.95,
  estimate_residual_method = "MoM",
  unmappable_effects = "inf",
  outlier.switch = T,
  Annotation = NULL,
  output.labels = NULL,
  carma.iter = 5,
  carma.inner.iter = 5,
  xQTL.max.num = 10,
  carma.epsilon.threshold = 0.001
)
}
\arguments{
\item{bX}{A matrix of effect estimates from the exposure GWAS.}

\item{bXse}{A matrix of standard errors of effect estimates from the exposure GWAS.}

\item{LD}{The linkage disequilibrium (LD) matrix.}

\item{xQTL.Nvec}{When choosing \code{"SuSiE"}, the vector of sample sizes of exposures.}

\item{xQTL.method}{The method used in purifying the xQTLs. SuSiE or CARMA can be used here, where the latter can be more accurate but much most computationally costly. Defaults is SuSiE.}

\item{xQTL.max.L}{When choosing \code{"SuSiE"}, the maximum number of L in estimating the xQTL effects. Defaults to 10.}

\item{xQTL.weight}{When choosing \code{"SuSiE"}, the vector of weights used in specifying the prior weights of SuSiE. Defaults to \code{NULL}.}

\item{xQTL.cred.thres}{When choosing \code{"SuSiE"}, the minimum empirical posterior inclusion probability (PIP) used in getting credible sets of xQTL selection. Defaults to \code{0.95}.}

\item{estimate_residual_method}{The method used for estimating residual variance. For the original SuSiE model, "MLE" and "MoM" estimation is equivalent, but for the infinitesimal model, "MoM" is more stable. We recommend using "Servin_Stephens" when n < 80 for improved coverage.}

\item{unmappable_effects}{The method for modeling unmappable effects: "none", "inf".}

\item{outlier.switch}{When choosing \code{"CARMA"}, an indicator of whether turning on outlier detection. Defaults to \code{F}.}

\item{Annotation}{When choosing \code{"CARMA"}, the annotation matrix of SNP. Default is NULL.}

\item{output.labels}{When choosing \code{"CARMA"}, output directory where output will be written while CARMA is running. Defaults to \code{NULL}, meaning that a temporary folder will be created and automatically deleted upon completion of the computation.}

\item{carma.iter}{When choosing \code{"CARMA"}, the maximum iterations for EM algorithm to run. Defaults to 5.}

\item{carma.inner.iter}{When choosing \code{"CARMA"}, the maximum iterations for Shotgun algorithm to run per iteration within EM algorithm. Defaults to 5.}

\item{xQTL.max.num}{When choosing \code{"CARMA"}, the maximum number of causal variants assumed per locus, which is similar to the number of single effects in SuSiE. Defaults to 10.}

\item{carma.epsilon.threshold}{When choosing \code{"CARMA"}, the convergence threshold measured by average of Bayes factors. Defaults to \code{1e-3}.}
}
\value{
A list containing the results of the MRBEEX analysis using different methods:
\describe{
  \item{\code{bXest}}{A matrix where each column represents \code{R * beta_PLS} for each exposure.}
  \item{\code{bXestse}}{A matrix where each column contains the standard errors of \code{R * beta_PLS} for each exposure.}
  \item{\code{bXest0}}{A matrix where each column represents \code{beta_PLS} for each exposure.}
  \item{\code{bXestse0}}{A matrix where each column contains the standard errors of \code{beta_PLS} for each exposure.}
  \item{\code{xQTLfitList}}{A list containing the xQTL selection results, which can be used in \code{CisMRBEEX} by setting \code{xQTLfitList = xQTLfitList}.}
}
}
\description{
This function performs infomative xQTL selection and sparse prediction using SuSiE and CARMA.
}
