% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_data.R
\name{summary_generation}
\alias{summary_generation}
\title{Generating simulated data for Mendelian randomization simulation}
\usage{
summary_generation(
  theta,
  m,
  Rbb,
  Ruv,
  Rnn,
  Nxy,
  Hxy,
  LD = "identify",
  non.zero.frac,
  pleiotropy.frac = 0,
  pleiotropy.var = 0.5,
  pleiotropy.cor = 0,
  pleiotropy.cor.exposure = c(1:length(theta)),
  pleiotropy.cor.effect = rep(1:length(theta)),
  effect.dis = "normal",
  pleiotropy.dis = "uniform"
)
}
\arguments{
\item{theta}{An (px1) vector of causal effects.}

\item{m}{The number of instrumental variables (IVs).}

\item{Rbb}{An (pxp) correlation matrix of genetic effects.}

\item{Ruv}{An ((p+1)x(p+1)) correlation matrix of residuals in outcome and exposures; the outcome is the first one.}

\item{Rnn}{An ((p+1)x(p+1)) correlation matrix of sample overlap; the outcome is the first one.}

\item{Nxy}{An ((p+1)x1) vector of GWAS sample sizes; the outcome is the first one.}

\item{Hxy}{An ((p+1)x1) vector of heritabilities; the outcome is the first one.}

\item{LD}{An (mxm) correlation matrix of the IVs or "identify" indicating independent IVs.}

\item{pleiotropy.frac}{A number in [0,0.5) indicating the fraction of IVs affected by pleiotropy.}

\item{pleiotropy.var}{A number in \eqn{[0, \infty)} indicating the variance attributed to pleiotropy.}

\item{pleiotropy.cor}{A number indicating the correlation between correlated horizontal pleiotropy and the specified latent variable. If set to 0, uncorrelated horizontal pleiotropy is generated.}

\item{pleiotropy.cor.exposure}{A vector of indices of variables that are correlated with the correlated horizontal pleiotropy.}

\item{pleiotropy.cor.effect}{A vector of effects corresponding to the variables correlated with the correlated horizontal pleiotropy. Specifically, \code{cor(bX[,pleiotropy.cor.exposure] \%*\% pleiotropy.cor.effect, CHP)} equals \code{pleiotropy.cor}.}

\item{effect.dis}{Distribution of genetic effects: "normal" (default), "uniform", or "t" distribution (with degree of freedom 5).}

\item{pleiotropy.dis}{Distribution of pleiotropy effects: "normal" (default), "uniform", or "t" distribution (with degree of freedom 5).}

\item{zero.frac}{An (px1) vector with all entries in (0,1]; each entry is the probability of deltaj such that betaj=betaj'*deltaj.}
}
\value{
A list containing simulated GWAS effect sizes for exposures (bX), their standard errors (bXse),
        the GWAS effect size for the outcome (by), its standard error (byse), the pleiotropy effects (pleiotropy), and the true effects.
}
\description{
This function generates simulated data for Mendelian Randomization (MR) analysis,
considering genetic effects, estimation errors, and horizontal pleiotropy.
It allows for different distributions of genetic effects and pleiotropy,
and accommodates both independent and correlated instrumental variables (IVs).
}
